% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-reliability.R
\name{cronbachs_alpha}
\alias{cronbachs_alpha}
\title{Estimate Cronbach's alpha coefficient of reliability}
\usage{
cronbachs_alpha(data, ...)
}
\arguments{
\item{data}{Name of the data frame containing the variables of interest.}

\item{...}{Names of variables (unquoted) on which to conduct analysis.}
}
\value{
Numeric estimate of Cronbach's alpha (0-1 scale)
}
\description{
Returns reliability estimate for a set of variables
}
\examples{

## estimate reliability of social media use items in polcom_survey data
cronbachs_alpha(polcom_survey, smuse1:smuse3)

}
