% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis.R
\name{factor_analysis}
\alias{factor_analysis}
\alias{eigenvalues}
\alias{parallel_analysis}
\title{Conduct factor analysis to estimate item loadings on each factor}
\usage{
factor_analysis(data, ..., n = NULL)

eigenvalues(data, ...)

parallel_analysis(data, ...)
}
\arguments{
\item{data}{Data frame of interest}

\item{...}{Variables (unquoted) of interest. If not provided, all variables
in data will be used}

\item{n}{Number of retained factors, defaulting to number of factors
suggested via parallel analysis.}
}
\value{
Output of factor analysis.

A tibble of p factors, their [eigen]values, and whether the values
  exceed 1.0
}
\description{
Provides estimates from a rotated factor analysis.

Calculates eigenvalues for p number of dimensions

Simulates data to use as a baseline for making factor/component retention
decisions
}
\examples{

## conduct parallel analysis of polcom news use items
factor_analysis(polcom, news_1:news_6)


## conduct parallel analysis of polcom news use items
eigenvalues(polcom, news_1:news_6)


## conduct parallel analysis of polcom news use items
parallel_analysis(polcom, news_1:news_6)

}
