% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fit.R
\name{nagelkerke}
\alias{nagelkerke}
\title{nagelkerke r squared}
\usage{
nagelkerke(m)
}
\arguments{
\item{m}{A GLM model object.}
}
\value{
R^2 estimate.
}
\description{
Estimate R^2 approximation for model object
}
\details{
Equation taken from the following study:
Nagelkerke, N. (1991). A Note on a General Definition of the Coefficient of Determination. Biometrika, 78(3), 691-692. doi:10.2307/2337038
}
