% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_regression.R
\name{ols_regression}
\alias{ols_regression}
\title{Ordinary least squares (OLS) regression}
\usage{
ols_regression(data, model, ...)
}
\arguments{
\item{data}{Data frame containing variables of interest.}

\item{model}{Model formula to be estimated.}

\item{...}{Other arguments passed to lm() function.}
}
\value{
A list object containing a "fit" and "ceof" data frames.
}
\description{
Conducts regression analysis to model outcome variable using OLS
}
\examples{

## predict mpg using weight and cylinders
ols_regression(datasets::mtcars, mpg ~ wt + cyl)

}
