% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-regression.R
\name{poisson_regression}
\alias{poisson_regression}
\title{Poisson regression for modeling of count data}
\usage{
poisson_regression(data, model, ...)
}
\arguments{
\item{data}{Data frame with variables of interest to model}

\item{model}{Specification of the theorized model}

\item{...}{Other args passed to glm}
}
\value{
List with fit and coef elements.
}
\description{
Conducts poisson regression (generalized linear models for count data)
}
