% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-regression.R
\name{ols_regression}
\alias{ols_regression}
\alias{logistic_regression}
\alias{poisson_regression}
\title{Ordinary least squares (OLS) regression}
\usage{
ols_regression(data, model, robust = FALSE, ...)

logistic_regression(data, model, robust = FALSE, ...)

poisson_regression(data, model, robust = FALSE, ...)

poisson_regression(data, model, robust = FALSE, ...)
}
\description{
Conducts regression analysis to model outcome variable using OLS

Conducts logistic regression analysis to model binary outcome variable using
  a generalized (binomial with logit link) linear model

Conducts negative binomial regression (generalized linear models for
overdispersed count data)

Conducts poisson regression (generalized linear models for count data)
}
