% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-regression.R
\name{regression_types}
\alias{regression_types}
\alias{ols_regression}
\alias{logistic_regression}
\alias{poisson_regression}
\alias{negbinom_regression}
\title{Types of regression models}
\usage{
ols_regression(data, model, robust = FALSE, ...)

logistic_regression(data, model, robust = FALSE, ...)

poisson_regression(data, model, robust = FALSE, ...)

negbinom_regression(data, model, robust = FALSE, ...)
}
\arguments{
\item{data}{Data frame containing variables in model}

\item{model}{Model formula to be estimated.}

\item{...}{Other arguments passed to modeling function.}
}
\description{
Conducts regression analysis to model outcome variable using OLS

Conducts logistic regression analysis to model binary outcome variable using
  a generalized (binomial with logit link) linear model

Conducts poisson regression (generalized linear models for count data)

Conducts negative binomial regression (generalized linear models for
overdispersed count data)
}
\details{
Available types of regression models.
}
