% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tbl_frame}
\alias{tbl_frame}
\title{Create a tibble data frame}
\usage{
tbl_frame(...)
}
\arguments{
\item{...}{A set of name-value pairs. Arguments are evaluated sequentially,
so you can refer to previously created variables.  These arguments are
processed with \code{\link[rlang:quos]{rlang::quos()}} and support unquote via \code{!!} and
unquote-splice via \code{!!!}.}
}
\value{
A tibble, class \code{c("tbl_df", "tbl", "data.frame")}
}
\description{
Create a tibble data frame
}
