% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-mlm.R
\name{tidy_mlm}
\alias{tidy_mlm}
\title{Multilevel model}
\usage{
tidy_mlm(.data, model, robust = FALSE, ...)
}
\arguments{
\item{robust}{Logial indicating whether to use a robust estimator}

\item{...}{Model formula to be estimated}

\item{data}{Data frame containing variables in model}
}
\value{
A model object
}
\description{
Conduct multilevel model analysis (AKA mixed models)
}
