% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-regression.R
\name{tidy_regression}
\alias{tidy_regression}
\title{Tidy regression analysis}
\usage{
tidy_regression(data, model, type = "ols", robust = FALSE, ...)
}
\arguments{
\item{data}{Data frame containing variables in model}

\item{model}{Model formula to be estimated.}

\item{type}{Type of regression model to use. Available types include
\code{"ols"}, \code{"logistic"}, \code{"poisson"}, \code{"negbinom"}. See
\link{regression_types} for more information.}

\item{robust}{Logical indicating whether to estimate a robust model. This
is available for all models but negative binomial.}

\item{...}{Other arguments passed to modeling function.}
}
\value{
A model object
}
\description{
Conducts regression analysis to model outcome variable using OLS, logistic,
poisson, and negative-binomial models
}
\details{
In addition to being a wrapper function for \link[base]{lm},
\link[base]{glm}, and robust models via \link[MASS]{rlm} (for robust OLS)
and \link[robust]{glmRob} this function (a) ensures \code{data} arguments
appear in first position for better consistency and easier piping and (b)
stores information about the call
}
\examples{

## predict mpg using weight and cylinders
m_ols <- datasets::mtcars \%>\%
  tidy_regression(mpg ~ wt + cyl)

## sweep and view summary
(s_ols <- sweep(m_ols))

## logistic regression predict am using disp, carb, and mpg
m_logistic <- datasets::mtcars \%>\%
  tidy_regression(am ~ disp + carb + mpg, type = "logistic")

## sweep and view summary
(s_logistic <- sweep(m_logistic))

## poisson regression predict cyl using disp, carb, and mpg
m_poisson <- datasets::mtcars \%>\%
  tidy_regression(cyl ~ disp + carb + mpg, type = "poisson")

## sweep and view summary
(s_poisson <- sweep(m_poisson))

}
