% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-sem.R
\name{tidy_sem}
\alias{tidy_sem}
\title{Structural equation model}
\usage{
tidy_sem(.data, ..., robust = FALSE)
}
\arguments{
\item{...}{Model formula to be estimated.}

\item{robust}{Logial indicating whether to use a robust estimator.}

\item{data}{Data frame containing variables in model}
}
\value{
A model object
}
\description{
Conduct latent variable model analysis (structural equation modeling)
}
