% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-sem.R
\name{tidy_sem_model}
\alias{tidy_sem_model}
\title{Build SEM model}
\usage{
tidy_sem_model(.data, ...)
}
\arguments{
\item{.data}{Data frame with variables of interest}

\item{...}{Model formulas following a slightly modified version of lavaan
model syntax (see \code{\link[lavaan]{lavParseModelString}}). Instead of
taking only a single character string representing the specified model
(as would be done with lavaan), this function accepts unquoted formulas
that use commas between variable/equation lines instead of line breaks
(note: line breaks may still be used for aesthetic/spacing purposes, but
they otherwise have no affect on how the model is processed).}
}
\value{
Returns the provided data with a "model" attribute
}
\description{
Specify a latent variable model, to be used prior to \code{\link{tidy_sem}}
}
