% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-factors.R
\name{factor_analysis}
\alias{factor_analysis}
\alias{eigenvalues}
\alias{parallel_analysis}
\title{Conduct factor analysis to estimate item loadings on each factor}
\usage{
factor_analysis(data, ..., nfactors = "parallel", rotation = "oblique",
  method = NULL)

eigenvalues(data, ...)

parallel_analysis(data, ...)
}
\arguments{
\item{data}{Data frame of interest}

\item{...}{Variables (unquoted) of interest. If not provided, all variables
in data will be used}

\item{nfactors}{Method used to estimate the number of factors. The default,
uses a "parallel" analysis to estimate the number of retained factors.
Also available is the "kaiser" method, which retains all eigenvalues
greater than 1.0. Users may also manually specify the number of factors
by entering a numeric [or integer] value here.}

\item{rotation}{Method used to estimate rotated solution. The default
"oblique" rotation method (geominQ from GPArotation pkg) assumes the
factors are related (a fairly typical or desirable assumption of social
science constructs). Also available is the "orthogonal" rotation method
(geominT from the GPArotation pkg), which assumes the factors are
independent. Users may override these options, providing their own
rotation method by specifying [presumably a method offered in the psych
or GPArotation packages] here as well.}
}
\value{
Output of factor analysis.

A tibble of p factors, their [eigen]values, and whether the values
  exceed 1.0
}
\description{
Provides estimates from a rotated factor analysis.

Calculates eigenvalues for p number of dimensions

Simulates data to use as a baseline for making factor/component retention
decisions
}
\examples{

## conduct parallel analysis of polcom news use items
factor_analysis(polcom, news_1:news_6)


## conduct parallel analysis of polcom news use items
eigenvalues(polcom, news_1:news_6)


## conduct parallel analysis of polcom news use items
parallel_analysis(polcom, news_1:news_6)

}
