% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{logistic_regression}
\alias{logistic_regression}
\title{Logistic regression for dichotomous outcomes}
\usage{
logistic_regression(data, model, ...)
}
\arguments{
\item{data}{Data frame containing variables of interest.}

\item{model}{Model formula to be estimated.}

\item{...}{Other arguments passed to lm() function.}
}
\value{
A list object containing a "fit" and "ceof" data frames.
}
\description{
Conducts logistic regression analysis to model binary outcome variable using
  a generalized (binomial with logit link) linear model
}
\examples{

## predict mpg using weight and cylinders
logistic_regression(datasets::mtcars, am ~ disp + carb + mpg)

}
