% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-regression.R
\name{ols_regression}
\alias{ols_regression}
\alias{ols_regression_robust}
\title{Ordinary least squares (OLS) regression}
\usage{
ols_regression(data, model, ...)

ols_regression_robust(data, model, ...)
}
\arguments{
\item{data}{Data frame containing variables of interest.}

\item{model}{Model formula to be estimated.}

\item{...}{Other arguments passed to lm() function.}
}
\value{
A list object containing a "fit" and "ceof" data frames.
}
\description{
Conducts regression analysis to model outcome variable using OLS
}
\examples{

## predict mpg using weight and cylinders
m1 <- ols_regression(datasets::mtcars, mpg ~ wt + cyl)

## sweep and view summary
(s1 <- sweep(m1))
}
